add stateless input and fix select styles
