+++ InputEmail/story.jsx
@@ -40,14 +40,4 @@ storiesOf('InputEmail', module)
         submitting: true,
       }}
     />
-  ))
-  .add('focused', () => (
-    <InputEmail input={input} focused />
-  ))
-  .add('onFocus + onBlur', () => (
-    <InputEmail
-      input={input}
-      onFocus={action('on-focus')}
-      onBlur={action('on-blur')}
-    />
   ));
