+++ InputPassword/index.jsx
@@ -1,26 +1,21 @@
 import React from 'react';
 import Input from '../Input';
+import InputStateless from '../InputStateless';
 
 const InputPassword = ({
   input,
   label,
   meta,
-  focused,
-  onFocus,
-  onBlur,
 }) =>
   <Input
     input={input}
     label={label}
     meta={meta}
     type={'password'}
-    focused={focused}
-    onFocus={onFocus}
-    onBlur={onBlur}
   />;
 
 InputPassword.propTypes = {
-  ...Input.commonPropTypes,
+  ...InputStateless.commonPropTypes,
 };
 
 export default InputPassword;
