+++ Select/index.jsx
@@ -4,10 +4,10 @@ import { calculateStyles } from '../lib/utils';
 import colors, {
   transparent,
   mystic,
-  shuttleGray,
 } from '../style/color';
 import {
-  fontSizeSmall,
+  fontSize,
+  fontFamily,
 } from '../style/font';
 import {
   borderRadius,
@@ -17,7 +17,7 @@ import {
 import ArrowDownIcon from '../Icon/Icons/ArrowDownIcon';
 import ArrowUpIcon from '../Icon/Icons/ArrowUpIcon';
 
-const height = 2;
+const height = 2.3;
 
 const defaultIconStyle = {
   position: 'absolute',
@@ -30,7 +30,7 @@ const selectWrapperStyle = {
   position: 'relative',
 };
 
-const RangeIcon = () =>
+const RangeIcon = () => (
   <div
     style={{
       position: 'absolute',
@@ -54,7 +54,7 @@ const RangeIcon = () =>
         fontFamily: 'none',
       }}
     >
-      <ArrowUpIcon size={'small'} />
+      <ArrowUpIcon size="small" />
     </div>
     <div
       style={{
@@ -62,14 +62,16 @@ const RangeIcon = () =>
         fontFamily: 'none',
       }}
     >
-      <ArrowDownIcon size={'small'} />
+      <ArrowDownIcon size="small" />
     </div>
-  </div>;
+  </div>
+);
 
-const DefaultIcon = () =>
+const DefaultIcon = () => (
   <div style={defaultIconStyle}>
     <ArrowDownIcon />
-  </div>;
+  </div>
+);
 
 /* eslint-disable react/prop-types */
 const SelectIcon = ({ noStyle, rangeSelector }) => {
@@ -101,8 +103,9 @@ const Select = ({
   const selectStyle = calculateStyles({
     default: {
       height: `${height}rem`,
-      padding: '0 1.5rem 0 0.5rem',
-      fontSize: fontSizeSmall,
+      padding: '0.5rem 1.5rem 0.5rem 0.5rem',
+      fontFamily,
+      fontSize,
       background: transparent,
       border: `${borderWidth} solid ${mystic}`,
       borderRadius,
@@ -111,6 +114,7 @@ const Select = ({
       MozAppearance: 'none',
       width: '100%',
       color: colors[color],
+      outline: 'none',
     },
     noStyle: {
       height: 'auto',
@@ -146,11 +150,11 @@ const Select = ({
         value={value}
       >
         {
-          options.map(option =>
+          options.map(option => (
             <option key={option.value.toString()} value={option.value}>
               {option.name}
-            </option>,
-            )
+            </option>
+          ))
         }
       </select>
       <SelectIcon
