+++ index.js
@@ -6,6 +6,7 @@ export HoverableText from './HoverableText';
 export Icon from './Icon';
 export IconArrowPopover from './IconArrowPopover';
 export Input from './Input';
+export InputStateless from './InputStateless';
 export InputAutocomplete from './InputAutocomplete';
 export InputDate from './InputDate';
 export InputEmail from './InputEmail';
